Move intermediateca to baseca
